requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-client-cache==0.7.3
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.1.0
babelfish==0.6.0
beautifulsoup4==4.11.1
charset-normalizer==2.1.1
ffmpeg-python==0.2.0
frozenlist==1.3.1
future==0.18.2
guessit==3.4.3
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
multidict==6.0.2
python-dateutil==2.8.2
python-forge==18.6.0
rebulk==3.1.0
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
url-normalize==1.4.3
webencodings==0.5.1
yarl==1.8.1
